% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_analyze_data_to_estimate.R
\name{analyze_plot_ts_all}
\alias{analyze_plot_ts_all}
\title{Iterates over US states and plots the enrgy raw counts and thresholds.}
\usage{
analyze_plot_ts_all(
  data,
  threshold_matrix,
  id_range = NULL,
  id_check,
  explode_here,
  explode_offset
)
}
\arguments{
\item{data}{data input}

\item{threshold_matrix}{the matrix of thresholds; numeric matrix of dimension
equal to the number of states (rows) and number of energy types (cols)
giving the threshold values for each energy component where to position the
dummy}

\item{id_range}{range of ids}

\item{id_check}{ids to check}

\item{explode_here}{point to explode}

\item{explode_offset}{magnitude of explode}
}
\value{
pure side effect functions with just plots
}
\description{
Iterates over US states and plots the enrgy raw counts and thresholds.
}
