% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_plot_data.R
\name{generate_all_plots}
\alias{generate_all_plots}
\title{Wrapper functions to generate all plot types}
\usage{
generate_all_plots(
  data_sources = NULL,
  save_data_sources = FALSE,
  data_prices = NULL,
  save_data_prices = FALSE,
  data_cumcap = NULL,
  save_data_cumcap = FALSE,
  plot_src_prc = TRUE,
  save_src_prc = FALSE,
  plot_src_ccap = TRUE,
  save_src_ccap = FALSE,
  path_to_save
)
}
\arguments{
\item{data_sources}{data set of shares as generated by
\code{generate_data_to_plot_shares}}

\item{save_data_sources}{logical; if \code{TRUE}, then resulting plots for
\code{data_sources} are saved}

\item{data_prices}{data set of shares as generated by
\code{generate_data_to_plot_price}}

\item{save_data_prices}{logical; if \code{TRUE}, then resulting plots for
\code{data_prices} are saved}

\item{data_cumcap}{dataset of cumulative generating capacities as
generated by \code{generate_data_to_plot_ccap}}

\item{save_data_cumcap}{logical; if \code{TRUE}, then resulting plots for
\code{plot_src_prc} are saved}

\item{plot_src_prc}{logical; if \code{TRUE}, type '03'-plots will be
returned that plot shares vs. prices with merged data via e.g.
\code{dplyr::full_join(data_shaes, data_prices)}}

\item{save_src_prc}{logical; if \code{TRUE}, then resulting plots for
\code{plot_src_prc} are saved}

\item{plot_src_ccap}{logical; if \code{TRUE}, type '04'-plots will be
returned that plot shares vs. cumcap with merged data via e.g.
\code{dplyr::full_join(data_shaes, data_cumcap)}}

\item{save_src_ccap}{logical; if \code{TRUE}, then resulting plots for
\code{plot_src_prc} are saved}

\item{path_to_save}{a character string giving the path where to save above
plots}
}
\value{
a list of 'ggplots' and, as a side effect, the resulting plots are
save too, if the corresponding \code{save_data_xxx= TRUE}
}
\description{
Wrapper functions to generate all plot types #'
}
