% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{generate_avail_vars}
\alias{generate_avail_vars}
\title{Function that generates availability dummy}
\usage{
generate_avail_vars(data, n_env)
}
\arguments{
\item{data}{current dataset as dataframe or tibble}

\item{n_env}{environment containing the names of variables of energy sources}
}
\value{
dataset as provided by the argument \code{data} but with additionally
the availability dummy variable
}
\description{
Availability dummy equals one if resource is >0, but is zero else
}
