% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{generate_constant}
\alias{generate_constant}
\title{Adds a constant to the dataset}
\usage{
generate_constant(data)
}
\arguments{
\item{data}{current dataset as dataframe or tibble}
}
\value{
dataset as before but with a column of "ones" (i.e. a constant)
}
\description{
Adds a constant to the dataset
}
