% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{generate_data_for_estimation}
\alias{generate_data_for_estimation}
\title{Generates data which is used for estimation.}
\usage{
generate_data_for_estimation(data, data_gdp, avail_thresholds)
}
\arguments{
\item{data}{either character string of the path to input data or the data
object itself in form of a tibble or data.frame}

\item{data_gdp}{either character string of the path to input data or the data
object itself, containing GDP data, in form of a tibble or data.frame}

\item{avail_thresholds}{a double matrix of dimension "us_states" x
"energy_types" i.e. it contains for every row (for every US state) as
columns/variables the corresponding energy threshold where the
\code{xxx_avail} variable is set from 0 to 1. So if the 3rd row and 2nd
column entry of this matrix is 1000, then, for the third US state which is
passed, and for the second energy type, the corresponding threshold is 1000
meaning that the dummy becomes =1 if the raw counts exceeds 1000, but is
zero otherwise.}
}
\value{
the resulting data with addtional variabls 'constant' the
\code{xxx_avail}s
}
\description{
The main additions to the raw data are:
\itemize{
\item{1.}{ a constant is added for all states and years}
\item{2.}{ the \code{xxx_avail} variable is set}
}
The latter varies  for all cross sectional units (all US states), and all
energy types!
}
