% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_generate_data_to_plot.R
\name{generate_data_to_plot_ccap}
\alias{generate_data_to_plot_ccap}
\title{Function that returns the states, years and prices of the data argument.}
\usage{
generate_data_to_plot_ccap(path_to_data, as_fractions = NULL)
}
\arguments{
\item{path_to_data}{character string giving the path to the data}

\item{as_fractions}{logical; if \code{TRUE}, then cumcaps are returened as
fractions rather than levels}
}
\value{
a data.frame/tibble that contains states, years and cumulative
generating capacities
}
\description{
Function that returns the states, years and prices of the data argument.
}
