% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_generate_data_to_plot.R
\name{generate_data_to_plot_prices}
\alias{generate_data_to_plot_prices}
\title{Function that returns the states, years and prices of the data argument.}
\usage{
generate_data_to_plot_prices(path_to_data)
}
\arguments{
\item{path_to_data}{character string giving the path to the data}
}
\value{
a data.frame/tibble that contains states, years and energy prices
}
\description{
Function that returns the states, years and prices of the data argument.
}
