% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_generate_data_to_plot.R
\name{generate_data_to_plot_shares}
\alias{generate_data_to_plot_shares}
\title{Function that computes shares of the energy types.}
\usage{
generate_data_to_plot_shares(path_to_data, as_fractions)
}
\arguments{
\item{path_to_data}{character string giving the path to the data}

\item{as_fractions}{logical; if \code{TRUE}, then shares are returned as
fractions rather than levels}
}
\value{
original dataset with additional variables:
\itemize{
\item{\code{CLEIB_share}:}{ share of coal}
\item{\code{NGEIB_share}:}{ share of natural gas}
\item{\code{PAEIB_share}:}{ share of petroleum}
\item{\code{NUEGB_share}:}{ share of nuclear power}
\item{\code{renewables_share}:}{ share of renewables}
}
where the latter may come in unaggregated form i.e. there are additional
variables:
\itemize{
\item{\code{GEEGB_share}:}{ share of geothermal}
\item{\code{SOEGB_share}:}{ share of soalr}
\item{\code{WYEGB_share}:}{ share of ?}
\item{\code{HYEGB_share}:}{ share of hydropower}
\item{\code{WWEIB_share}:}{ share of hydropower}
}
}
\description{
Function that computes shares of the energy types.
}
