% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{generate_other_vars}
\alias{generate_other_vars}
\title{Adds other variables to the dataset}
\usage{
generate_other_vars(data, ...)
}
\arguments{
\item{data}{the current dataset where to add additional variables}

\item{...}{additional datasets with cross sectional identifier \code{state}
and year identifier \code{year} to be merged (\code{dplyr::left_join()}),
and remaining columns being the variables to add to the \code{data} dataset}
}
\value{
data with shares for energy sources, avail-dummies and cumulative
capacities in the "correct" variable order
}
\description{
Adds other variables to the dataset
}
