% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_plot_data.R
\name{generate_plot_ts}
\alias{generate_plot_ts}
\title{Function that generates a single time series plot.}
\usage{
generate_plot_ts(data_ts, name_state, names_ts, names_to_display)
}
\arguments{
\item{data_ts}{tibble/data.frame of energy types (i.e. shares and/or raw
counts) or energy prices}

\item{name_state}{character string of the name of the state for which to plot}

\item{names_ts}{names of the time series variables to plot}

\item{names_to_display}{how the names should occur in the final plots (labels
etc. typically energy type names or names of the prices)}
}
\value{
a plot of the data: either 3x2 (for shares or raw counts) and a plot
on top; for prices a 3x1 plot
}
\description{
The time series can be either the energy types directly i.e. their shares or
the raw counts, or, the energy prices.
}
