% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_plot_data.R
\name{generate_plot_ts_src_ccap}
\alias{generate_plot_ts_src_ccap}
\title{Function that generates two time series plot.}
\usage{
generate_plot_ts_src_ccap(data_ts, name_state, names_ts, names_to_display)
}
\arguments{
\item{data_ts}{tibble/data.frame of energy types (i.e. shares and/or raw
counts) or energy prices}

\item{name_state}{character string of the name of the state for which to plot}

\item{names_ts}{list: names of the time series variables to plot (list of
two)}

\item{names_to_display}{how the names should occur in the final plots (labels
etc. typically energy type names or names of the prices)}
}
\value{
a plot of the data in 3x2-form containing the shares or raw counts,
the corresponding time series of prices (if available) and a plot on top
}
\description{
The time series are those of shares and cumulative capacities.
}
