% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{generate_shares}
\alias{generate_shares}
\title{Adds shares to the data as well as avail-dummies and cumulative capacities}
\usage{
generate_shares(data, n_env)
}
\arguments{
\item{data}{the dataset for which to compute shares, and add avail dummies,
cummulative capacities and avail x cummulative capacities}

\item{n_env}{environment containing the names of variables of energy
sources}
}
\value{
data with shares for energy sources, avail-dummies and cumulative
capacities in the "correct" variable order
}
\description{
Adds shares to the data as well as avail-dummies and cumulative capacities
}
