% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_GET.R
\name{get_clean_data_gdp}
\alias{get_clean_data_gdp}
\title{Generates tidy GDP data in long format from a raw (untidy) dataset}
\usage{
get_clean_data_gdp(path_to_data)
}
\arguments{
\item{path_to_data}{character string specifying path to GDP data without
trailing '/'}
}
\value{
a tibble/dataframe of GDP data in long format
}
\description{
Generates tidy GDP data in long format from a raw (untidy) dataset
}
