% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_GET.R
\name{get_cumcap_precleaned}
\alias{get_cumcap_precleaned}
\title{Preleaning raw cumulative generating capacities data}
\usage{
get_cumcap_precleaned(path_to_data)
}
\arguments{
\item{path_to_data}{character string containing the path to the raw
cumulative generating capacities data}
}
\value{
list of two; precleaned cumulative capacities data with first element
being the operational power plant dataset and the second the one containing
the retired power plants
}
\description{
Preleaning raw cumulative generating capacities data
}
