% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_helper_prepare_panel_data.R
\name{get_data}
\alias{get_data}
\title{Retrieves necessary subset of the data to be used for subsequent estimations.}
\usage{
get_data(data_full, states_to_use, Z_to_use, U_to_use)
}
\arguments{
\item{data_full}{full dataset (as a dataframe or tibble)}

\item{states_to_use}{character vector of names for cross sectional units
(i.e. US state names) to conisder}

\item{Z_to_use}{a list of dimension equel to the number of dirichlet
shares/fractions i.e. with each component being a vector of regressor names
for the Z-type regressors (regressors for which parameters/coefficients are
shared among all US-states)}

\item{U_to_use}{a list of dimension equel to the number of dirichlet
shares/fractions i.e. with each component being a vector of regressor names
for the U-type regressors (regressors for which parameters/coefficients
vary among all US-states)}
}
\value{
a list with 6 components:
\itemize{
\item{y_t}{: an array of dimension \code{TTxDDxNN} (time series length,
number of components, and number of cross sectional units) containing
the measurements e.g. the raw energy counts or dirichlet shares}
\item{num_counts:}{a matrix of dimension \code{TTxNN}; the second part of
the measurements (for some distributions necessary e.g. the dirichlet
multinomial) e.g. the number of total energy counts per time and per
cross sectional unit}
\item{Z:}{ an array of dimension
\code{TTx(DD*num_regressors_per_component)xNN} containing the Z-type
regressors; for each component \code{DD}; there might be a different
number of regressors used so the number of columns has the component}
\code{num_regressors_per_component}; there are \code{NN} different cross
sectional units (US states), so the final array dimension is \code{NN}
\item{U:}{ an array of dimension
\code{TTx(DD*num_regressors_per_component)xNN} containing the U-type
regressors; for each component \code{DD}; there might be a different
number of regressors used so the number of columns has the component}
\code{num_regressors_per_component}; there are \code{NN} different cross
sectional units (US states), so the final array dimension is \code{NN}
\item{NN:}{number of cross sectional units (US states)}
\item{TT:}{number of observed time periods (e.g. 1960-2014)}
\item{DD:}{number of energy components (corresponding to dirichlet shares
or raw counts of energy as e.g. with the dirichlet-multinomial)}
}
}
\description{
Retrieves necessary subset of the data to be used for subsequent estimations.
}
