% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_GET.R
\name{get_data_eia}
\alias{get_data_eia}
\title{Get data from the EIA webpage}
\usage{
get_data_eia(key = NULL)
}
\arguments{
\item{key}{key to access EIA web API via R (can be requested from the
webpage}
}
\value{
a dataset
}
\description{
Get data from the EIA webpage
}
