% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{get_final_selection}
\alias{get_final_selection}
\title{Does a final selection on the dataset to get only relevant variables}
\usage{
get_final_selection(data, n_env)
}
\arguments{
\item{data}{current dataset as dataframe or tibble}

\item{n_env}{environment containing the names of variables of energy sources}
}
\value{
dataset as provided by the argument \code{data} but with only
relevant variables selected (see function body)
}
\description{
Does a final selection on the dataset to get only relevant variables
}
