% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_helper_prepare_panel_inits.R
\name{get_inits_from_prior_output}
\alias{get_inits_from_prior_output}
\title{Helper function to obtain initial values from prior estimations.}
\usage{
get_inits_from_prior_output(
  path_to_inits,
  dim_reg_z = NULL,
  dim_reg_u = NULL,
  reg_names_Z = NULL,
  reg_names_U = NULL
)
}
\arguments{
\item{path_to_inits}{path to \code{.RData}-file of the output of
\code{pgas_xxx_R()} or \code{pgas_xxx_cpp()}}

\item{dim_reg_z}{either \code{NULL} or the vector giving the z-type regressor
dimensions (typically used when working with simulated data where there are
no regressor names); if \code{NULL}, then \code{reg_names_z} must be
specified!}

\item{dim_reg_u}{either \code{NULL} or the vector giving the u-type regressor
dimensions (typically used when working with simulated data where there are
no regressor names); if \code{NULL}, then \code{reg_names_} must be
specified!}

\item{reg_names_Z}{either \code{NULL} or the list of character vectors
giving the z-type regressor names (typically used when working with real
data); if \code{NULL}, then \code{dim_reg_z} must be specified!}

\item{reg_names_U}{either \code{NULL} or the list of character vectors
giving the u-type regressor names (typically used when working with real
data); if \code{NULL}, then \code{dim_reg_u} must be specified!}
}
\value{
a list of two: \code{par_inits} and \code{states_init} with
structures to be directly used as an argument to \code{pgas_xxx_R()} or
\code{pgas_xxx_cpp()} type function
}
\description{
The output of the \code{pgas_xxx_R()} or \code{pgas_xxx_cpp()} is used to
obtain initial starting/initial values for parameters and latent states for
subsequent estimation. The last (particle) MCMC iteration is used.
}
