% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{get_names_vars}
\alias{get_names_vars}
\title{Adjust negative values for observed (megawatt) counts.}
\usage{
get_names_vars(env)
}
\arguments{
\item{env}{environment containing the names of variables of energy
sources}
}
\value{
invisible; plain side effect function assigning variables to caller
environment
}
\description{
Adjust negative values for observed (megawatt) counts.
}
