% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_helper_prepare_panel_data.R
\name{get_state_names}
\alias{get_state_names}
\title{Helper function to retrieve US state names to be used.}
\usage{
get_state_names(data, num_biggest)
}
\arguments{
\item{data}{the dataset containing all states and data}

\item{num_biggest}{numeric index vector specifiying the indices (first to
last ) number of states names to retrieve, starting with the most
populated. So \code{num_biggest = 1:10} returns the names of the first 10
biggest US states (according to population).}
}
\value{
a character vector of states that satisfies above conditions.
}
\description{
First, states that have a-typical behaviour are excluded: US (which is the
overall USA so actually no states), then states that have permanent zeros
(never produce a particular type of energy), then 'NJ' (which is just weird)
and finally CO (just like that, no particular reason). The remaining states
are retrieved in order of their population, see \code{num_biggest} for
details.
}
