% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_helper_prepare_panel_data.R
\name{get_states_init}
\alias{get_states_init}
\title{Returns initialization for latent states}
\usage{
get_states_init(y_t, NN, TT, DD)
}
\arguments{
\item{y_t}{the measurements e.g. for the dirichlet multinomial the raw energy
counts}

\item{NN}{cross sectional dimension}

\item{TT}{time series dimension}

\item{DD}{number of components in the measurement vector
(dimension of \code{y_t} i.e. number of energy types)}
}
\value{
an array of dimension \code{TTxDDxNN} containing the initializiation
values for the latent states
}
\description{
This is required for the SMC/particle filter; the current method infers
starting values for the states from the data: zeros are set to a small value
termed 'zero_lower_bound', while the other state components are proportional
to the log of the scaled measurement y_t of the corresponding energy
fraction (the \code{d} component). Scaling of the \code{y_t[t, d, n]} is
done with a scaling factor before taking logs (by default set to 1).
}
