% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_generate_data_to_estimate_COMPUTE.R
\name{merge_into_renewables}
\alias{merge_into_renewables}
\title{Merging components of renewable sources into single category}
\usage{
merge_into_renewables(data, components_to_merge = NULL, cumcap_to_merge = NULL)
}
\arguments{
\item{data}{a dataframe or tibble containing all resource categories}

\item{components_to_merge}{a character string of (variable) names for energy
sources (shares) corresponding to col-names of data which will be merged
into the new "renewables" category}

\item{cumcap_to_merge}{a character string of cumulative-capacity (variable)
names corresponding to col-names of data which will be merged into the new
"renewables" category}
}
\value{
a dataframe or tibble; same as argument data but with renewable
sources grouped into a single category, removed columns of
'components_to_merge', and adjusted cumulative generating capacities.
}
\description{
Merging components of renewable sources into single category called
'renewables'.
}
