% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USenergyDataCleaning.R
\docType{package}
\name{USenergyDataCleaning}
\alias{USenergyDataCleaning}
\title{The data wrangling package for the usa_energy project}
\description{
Mainly plotting, descriptive visualization and data postprocessing stuff
related to this project. For internal use only.
}
