% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_analyze_data_to_estimate.R
\name{view_problematic}
\alias{view_problematic}
\title{Return the problematic cases for raw or final data sets}
\usage{
view_problematic(
  data,
  negatives,
  zeros,
  missing_raw,
  missing_shares,
  cumcap_wrong
)
}
\arguments{
\item{data}{a data frame or tibble with raw data}

\item{negatives}{logical; if \code{TRUE}, negative cases are returned}

\item{zeros}{logical; if \code{TRUE}, zero cases are returned}

\item{missing_raw}{logical; if \code{TRUE}, missing values for raw
consumption units are returned}

\item{missing_shares}{logical; if \code{TRUE},  missing values for shares are
returned}

\item{cumcap_wrong}{logical; if \code{TRUE}, cases where cumulative
capacities are zero, though the corresponding resource is consumed are
returned}
}
\value{
A named list with elements 'negatives', 'zeros', 'missing_raw',
'missing_shares', 'cumcap_wrong'.
}
\description{
Return problematic cases for raw or final data sets for analysis purposes.
Helps to justify if shares should be merged (e.g. to merge hydropower into a
component of renewable sources).
}
\details{
Possible problematic cases for raw or final data sets can be:
\itemize{
\item{negatives: negative values for energy consumption not allowed}
\item{zeros: zero values for energy consumption possible, but annoying}
\item{missing values for raw consumption units}
\item{missing values for shares of energy consumption (mix): occurs e.g. if
divided by a sum of energy consumption which is zero (nothing produced,
recorded)}
\item{cases where cumulative capacities are zero, though the corresponding
resource is consumed are returned}
}
}
